+++ sdk/webpubsub/azure-messaging-webpubsub/src/samples/java/com/azure/messaging/webpubsub/BroadcastingSample.java
@@ -3,34 +3,31 @@
 
 package com.azure.messaging.webpubsub;
 
-import com.azure.core.http.rest.RequestOptions;
-import com.azure.core.util.BinaryData;
 import com.azure.core.util.Configuration;
-import com.azure.core.util.Context;
+import com.azure.messaging.webpubsub.models.WebPubSubContentType;
 
 public class BroadcastingSample {
     private static final String CONNECTION_STRING = Configuration.getGlobalConfiguration().get("WEB_PUB_SUB_CS");
 
     public static void main(String[] args) {
         // build a sync client
-        WebPubSubServiceClient chatHub = new WebPubSubServiceClientBuilder()
+        WebPubSubServiceClient chatHub = new WebPubSubClientBuilder()
             .connectionString(CONNECTION_STRING)
             .hub("chat")
             .buildClient();
 
         // send a text message to the entire hub
-        chatHub.sendToAll("{\"message\": \"Hello world!\"}", "application/json");
+        chatHub.sendToAll("{\"message\": \"Hello world!\"}");
 
         // send a text message to a particular group
-        chatHub.sendToGroup("admin", "Hi admins!", "text/plain");
+        WebPubSubGroup adminGroup = chatHub.getGroup("admin");
+        adminGroup.sendToAll("Hi admins!", WebPubSubContentType.TEXT_PLAIN);
 
         // send binary data to the entire hub
         byte[] data = new byte[10];
         for (int i = 0; i < 10; i++) {
             data[i] = (byte) i;
         }
-        chatHub.sendToAllWithResponse(BinaryData.fromBytes(data), new RequestOptions()
-                .addRequestCallback(request -> request.getHeaders().set("Content-Type", "application/octet-stream")),
-                Context.NONE);
+        chatHub.sendToAll(data);
     }
 }