+++ sdk/webpubsub/azure-messaging-webpubsub/CHANGELOG.md
@@ -10,7 +10,7 @@
 - Upgraded `azure-core-http-netty` to `1.10.1`.
 
 ### Breaking Changes
-- Changed sync and async clients to use protocol methods that uses `RequestOptions` and `BinaryData` to create HTTP 
+- Changed sync and async clients to use protocol methods that uses `RequestOptions` and `BinaryData` to create HTTP
   request.
 
 ## 1.0.0-beta.2 (2021-04-27)